requirements.txt
Flask==2.0.3
Flask-Cors==3.1.1
requests==2.26.0
python-dotenv==0.19.2
apscheduler==3.7.0
